+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webkit_constants.dart
@@ -126,4 +126,9 @@ class WebViewFeatureConstants {
   ///
   /// See https://developer.android.com/reference/androidx/webkit/WebViewFeature#PAYMENT_REQUEST.
   static const String paymentRequest = 'PAYMENT_REQUEST';
+
+  /// This feature covers [WebViewCompat.addDocumentStartJavaScript].
+  ///
+  /// See https://developer.android.com/reference/androidx/webkit/WebViewFeature#DOCUMENT_START_SCRIPT.
+  static const String documentStartScript = 'DOCUMENT_START_SCRIPT';
 }